Split tags into specific and union ones
